requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
attrs==25.3.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
discord.py==2.5.2
python-dotenv==1.1.1
requests==2.32.5
six==1.17.0
urllib3==2.5.0
flask==3.1.2
numpy==1.26.4
pandas==2.3.3
matplotlib==3.10.7
